Tweak diff engine to properly support: - accents - email addresses and single words - escaped HTML including &lt; and &gt;
More info: Previously the statemachine has 4 states, :tag, :char, :whitespace, and indeterminant (which meant continue with the previous state), as certain chars where neither :char nor :whitesapce. I removed the indeterminant case so now a char is either a :wordchar (in which case join to the word being parsed) or an :other in which case go to the next word.